+++ test/resources/ns_with_multiple_shorthand_meta.clj
@@ -1,0 +1,2 @@
+(ns ^:automation ^:multiple ns-with-multiple-shorthand-meta
+  (:require [clojure.string]))